requirements.txt
asgiref==3.8.1
Django==5.1.7
django-filter==25.1
django-seed==0.3.1
Faker==37.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
toposort==1.10
tzdata==2025.1
